Rename: main -> subsearch
